Move DomainWorkflows to root of view
